Fix: handle type casting (int) properly

+++ mp2/src/sender_main.c
@@ -203,7 +203,7 @@ int rdt_sender_event_timeout(rdt_sender_ctrl_info_t *ctrl, char *sendbuf) {
     ctrl->state = SS;
 
     rdt_sender_act_retransmit(ctrl, sendbuf,
-                              max(0, (int)ctrl->expack - DATA_LEN));
+                              max(0, (int)(ctrl->expack - DATA_LEN)));
     return 1;
 }
 
@@ -283,12 +283,12 @@ int rdt_sender_event_handleack(rdt_sender_ctrl_info_t *ctrl,
                 case FR:
                     ctrl->seq = recvpkt->header.ack;
                     ctrl->bytes_remaining -=
-                        min(DATA_LEN, ctrl->bytes_remaining);
+                        min((int)DATA_LEN, ctrl->bytes_remaining);
                     ctrl->cwnd = ctrl->ssthresh;
                     ctrl->dupack = 0;
                     ctrl->state = CA;
                     ctrl->expack = recvpkt->header.ack
-                        + min(DATA_LEN, ctrl->bytes_remaining);
+                        + min((int)DATA_LEN, ctrl->bytes_remaining);
                     break;
 
                 default:
